Fixed nuclear integral incorrect values; Was adding to an uninitialised Eigen Array so was getting funky values; initialising to zero worked

+++ src/BasisFunc.cpp
@@ -41,7 +41,7 @@ BasisFunction::BasisFunction(std::array<int, 3> a,
  */
 void BasisFunction::normalize(){
   int l = shell[0]; int m = shell[1]; int n = shell[2];
-  double L = l+m+n;
+  int L = l+m+n;
   for (int i = 0; i < exps.size(); i++){
     double num = std::pow(2.0, 2.0*(l+m+n) + 1.5) * std::pow(exps[i], (l+m+n) + 1.5);
     double denom = double_factorial(2*l-1) * double_factorial(2*m-1) * double_factorial(2*n-1) * std::pow(M_PI, 1.5);
